requirements.txt
asgiref==3.9.2
Django==5.2.7
sqlparse==0.5.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.5.0
django-filter==24.3
